intel-scraper/requirements.txt
# Supabase client
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client for proxy rotation
httpx>=0.27.0

# Playwright for browser automation
playwright>=1.40.0
playwright-stealth>=1.0.6


